backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic-settings==2.3.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pandas
openpyxl
openai
